+++ src/epoch/gregorian.rs
@@ -270,7 +270,7 @@ impl Epoch {
                 }
                 Some(days) => {
                     // Initialize the duration as the number of days since the reference year (may be negative).
-                    Unit::Day * i64::from(days)
+                    Unit::Day * i128::from(days)
                 }
             },
         };
@@ -301,13 +301,13 @@ impl Epoch {
         };
 
         // Add the number of days based on the input month
-        duration_wrt_ref += Unit::Day * i64::from(cumul_days[(month - 1) as usize]);
+        duration_wrt_ref += Unit::Day * i128::from(cumul_days[(month - 1) as usize]);
         // Add the number of days based on the input day and time.
-        duration_wrt_ref += Unit::Day * i64::from(day - 1)
-            + Unit::Hour * i64::from(hour)
-            + Unit::Minute * i64::from(minute)
-            + Unit::Second * i64::from(second)
-            + Unit::Nanosecond * i64::from(nanos);
+        duration_wrt_ref += Unit::Day * i128::from(day - 1)
+            + Unit::Hour * i128::from(hour)
+            + Unit::Minute * i128::from(minute)
+            + Unit::Second * i128::from(second)
+            + Unit::Nanosecond * i128::from(nanos);
 
         if second == 60 {
             // Herein lies the whole ambiguity of leap seconds. Two different UTC dates exist at the
@@ -624,9 +624,9 @@ impl Epoch {
 
         let tz = if offset_sign > 0 {
             // We oppose the sign in the string to undo the offset
-            -(i64::from(decomposed[7]) * Unit::Hour + i64::from(decomposed[8]) * Unit::Minute)
+            -(i128::from(decomposed[7]) * Unit::Hour + i128::from(decomposed[8]) * Unit::Minute)
         } else {
-            i64::from(decomposed[7]) * Unit::Hour + i64::from(decomposed[8]) * Unit::Minute
+            i128::from(decomposed[7]) * Unit::Hour + i128::from(decomposed[8]) * Unit::Minute
         };
 
         let epoch = Self::maybe_from_gregorian(
